site/backend/requirements.txt
fastapi
uvicorn
h11
httpx
google-genai
python-multipart
jinja2
gunicorn
reportlab
pandas
slowapi
python-dotenv
email-validator
cffi
xlrd
xlwt
beautifulsoup4
lxml
openpyxl
Pillow
APScheduler
idna
charset-normalizer
certifi
urllib3
psycopg2-binary
python-docx
aiosmtplib
redis
